+++ app/(dashboard)/layout.tsx
@@ -4,6 +4,7 @@ import React from 'react';
 import { ThemeBackground } from '@/components/ThemeBackground';
 import { ThemeSelector } from '@/components/ThemeSelector';
 import { AuthGuard } from '@/components/auth/AuthGuard';
+import { AuthProvider } from '@/components/providers/AuthProvider';
 import { useAuth } from '@/hooks/useAuth';
 import { useTimerTitle } from '@/hooks/useTimerTitle';
 import { useActiveSessionValidator } from '@/hooks/useActiveSessionValidator';
@@ -13,20 +14,18 @@ import { ToastContainer } from '@/components/ui/ToastContainer';
 import { AppBar, Toolbar, Typography, Box, Link as MUILink } from '@mui/material';
 import Link from 'next/link';
 
-export default function DashboardLayout({ children }: { children: React.ReactNode }) {
+function DashboardContent({ children }: { children: React.ReactNode }) {
   const { user, logout } = useAuth();
   const { activeSession, setActiveSession } = useTaskStore();
 
-  useTimerTitle(); // タイマータイトル更新を有効化
+  useTimerTitle();
 
-  // localStorage内のactiveSessionをFirestoreと照合し、不一致の場合はクリア
   useActiveSessionValidator(user?.id, activeSession, setActiveSession);
 
   return (
     <AuthGuard>
       <ToastContainer />
       <Box sx={{ minHeight: '100vh', bgcolor: 'grey.50', position: 'relative' }}>
-        {/* テーマ背景 */}
         <ThemeBackground />
         <AppBar
           position="static"
@@ -124,5 +123,13 @@ export default function DashboardLayout({ children }: { children: React.ReactNod
         </Box>
       </Box>
     </AuthGuard>
+  );
+}
+
+export default function DashboardLayout({ children }: { children: React.ReactNode }) {
+  return (
+    <AuthProvider>
+      <DashboardContent>{children}</DashboardContent>
+    </AuthProvider>
   );
 }
